fix: profile controller

+++ app/Http/Controllers/ProfileController.php
@@ -14,11 +14,17 @@ use App\Models\User;
 
 class ProfileController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        $this->middleware('verified');
+        $this->middleware('status');
+    }
+    
     public function show($slug)
     {
         $user = User::where('slug', $slug)->firstOrFail();
-        $posts = $user->posts;
-        return view('profile/profile', compact('user', 'posts'));
+        return view('profile/profile', compact('user'));
     }
 
     public function edit($slug)
@@ -29,29 +35,53 @@ class ProfileController extends Controller
 
     public function update(Request $request, $id)
     {
-        $image = $request->file('image');
-        $dateTime = Carbon::now();
-        $name = $dateTime->format('Ymd_His') . '.webp';
-        Image::make($image)->encode('webp', 75)->save(storage_path('app/public/images/profiles/' . $name));
+        $request->validate([
+            'image' => 'nullable|image',
+        ], [
+            'image.image' => 'El archivo debe ser una imagen.',
+        ]);
 
         $user = User::find($id);
-        $user->name = $request->name;
-        $user->slug = str::slug($request->name);
-        $user->description = $request->description;
 
-        $img = $user->image;
-        Storage::delete('public/images/profiles/' . $img);
-        $user->image = $name;
+        if($request->name != null){
+            $user->name = $request->name;
+            $user->slug = str::slug($request->name);
+        }
+
+        if($request->description != null){
+            $user->description = $request->description;
+        }
+
+        if($request->hasFile('image') && $request->file('image')->isValid()){
+            $img = $user->image;
+            Storage::delete('public/images/profiles/' . $img);
+
+            $image = $request->file('image');
+            $dateTime = Carbon::now();
+            $name = $dateTime->format('Ymd_His') . '.webp';
+            Image::make($image)->encode('webp', 75)->save(storage_path('app/public/images/profiles/' . $name));
+            $user->image = $name;
+        }
 
         $user->save();
 
         return Redirect::route('profile.show', ['slug' => $user->slug]);
     }
 
+    public function destroy($id)
+    {
+        $user = User::find($id);
+        $user->status = 0;
+        $user->save();
+
+        $img = $user->image;
+        Storage::delete('public/images/profiles/' . $img);
+        return redirect("/home");
+    }
+
     public function editPassword($slug)
     {
-        $user = User::where('slug', $slug)->firstOrFail();
-        return view('profile/password', compact('user'));
+        return view('profile/password');
     }
 
     public function updatePassword(Request $request)
@@ -59,6 +89,11 @@ class ProfileController extends Controller
         $request->validate([
             'old_password' => 'required',
             'password' => 'required|min:8|confirmed',
+        ],[
+            'old_password.required' => 'La contraseña antigua es requerida.',
+            'password.required' => 'La contraseña es requerida.',
+            'password.min' => 'La contraseña debe tener al menos 8 caracteres.',
+            'password.confirmed' => 'Las contraseñas no coinciden.',
         ]);
 
         $user = Auth::user();
@@ -76,17 +111,7 @@ class ProfileController extends Controller
 
     public function posts($slug)
     {
-        $user = User::where('slug', $slug)->firstOrFail();
-        $posts = $user->posts;
-        return view('profile/posts', compact('user', 'posts'));
-    }
-
-    public function destroy($id)
-    {
-        $user = User::find($id);
-        $user->delete();
-        $img = $user->image;
-        Storage::delete('public/images/profiles/' . $img);
-        return redirect("/home");
+        $user = User::with('posts')->where('slug', $slug)->firstOrFail();
+        return view('profile/posts', compact('user'));
     }
 }
